filter: faiss.search_presigned does not support id selector in python
To search on specific cluster, we need to provide the coarse_dis,
though it seems passing all zero still works because we are using
IVFFlat, not PQ etc.

Within a cluster, we still need the ID selector because having some
vectors in the clsuter matching the attributes filter does not mean
all the vectors within the cluster match the filter.

But faiss's python interface does not support passing search parameters
for search_preassigned